requirements.txt
discord.py
python-dotenv
requests
google-generativeai>=0.8.0 